Populate otp secrets when calling generators (#7)
OTP secrets are populated by the controller. It is more intuitive to
ensure secrets are present when calling otp generators when using those
methods directly in the application. So that end user does not have to
remember to populate the secrets manually.

Don't memoize otp generators.

Make otp_authenticator_qrcode availabe in the model.

---------

Co-authored-by: Hassan Eslaminezhad (Salari) <75081335+hassaneslami2765@users.noreply.github.com>